% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_api.R
\name{stringdist_api}
\alias{stringdist_api}
\title{Calling stringdist from \code{C} or \code{C++}}
\description{
As of version \code{0.9.5.0} several \code{C} level functions can be linked to
and called from C code in other R packages.

A description of the API can be found in \href{../doc/stringdist_api.pdf}{stringdist_api.pdf}.
}
